requirements.txt
sseclient-py
requests
flask